requirements.txt
python-docx>=1.0.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
transformers 
torch
keybert
lancedb
pandas 
pyarrow 
fastapi
chromadb>=0.4.0
sentence-transformers
scikit-learn
pypdf2
python-dotenv>=1.0.0